Got more stuff working
Got the Create Agent screen working.
Got the Create Agent button working.
Got the Cancel button in the Create Agent screen working.
Known Bug:
Clicking on Create Agent will lock the game up which will force you to
do a hard close of the game. It is somewhere in the Save portion of the
coding.